[plan-2026-08-19-499b6f0e/iter-2/step-18.1] writing 53 precision arms found five real fp16 defects and reversed two of its own readings
R-088/R-141 adoption for the 53 packages step 18 left open. The arm is
parametrized over one subject registry rather than copied 51 times, because
the assertions are identical and only the build/input pair differs.

Six subjects were NOT green. Five were product defects, fixed at source:

  SAM         float32 pixel_mean/pixel_std met a half tensor -- every
              mixed_float16 forward raised inside preprocess          D-063
  layers/moe  ops.one_hot returns float32 regardless of policy, so the
              routing mask met a float16 expert output; found via qwen,
              carried by EVERY hard-routed MoE consumer               D-064
  som         Keras autocasts a float32 variable read inside call, so
              the decayed sigma met the non-autocast float32 grid --
              training=True raised, training=False was green          D-062
  shgcn       ops.where differentiates the branch it DISCARDS; at the
              origin tanh(u)/u's 1/u**2 underflows in half precision,
              so curvature_theta's gradient was nan with the forward
              finite in both arms                                     D-061
  superpoint  TF registers no float16 ResizeBicubic gradient and
              returns None SILENTLY -- the whole descriptor head got
              no update under mixed precision, 2 None of 183          D-060

The sixth, yolo12's 5712-of-5712 NaN, is NOT an fp16 defect: its untrained
BatchNorms leave the float32 inference activations at absmax 2.997772e+08,
above float16's 65504 ceiling. 60 warm-up steps did not clear it and 200 did
(3.332934e+02), but judging in training mode costs 7.3s against 106.8s and
measures the regime fit() uses.                                       D-065

Two readings were reversed by the instrument's own controls. mobilenet and
squeezenet measured grad_norm_sum EXACTLY 0.0 -- the gradient of a symmetric
loss on a uniform softmax is zero by symmetry, not a dead model; the loss is
now ramp-weighted.                                                    D-059
And part 2 convicted BERT, FNet and SOM for returning an int32 mask beside
their float outputs; it now ranges over the float leaves, with a vacuity
guard so an all-integer output still fails.                           D-058

Every relaxation is proved RED against the defect it was carved around, and
every source fix is pinned in its own package's test_precision_arm.py against
the pre-fix expression re-created verbatim.

Measured green: 54/54 family, 22/22 oracle RED proofs, 510 passed 1 skipped
across the five touched model packages, 196/196 across the touched layer
tests, 509 passed 3 skipped test_package_api_contract, 67/67 repo-map numbers.

+++ src/dl_techniques/layers/moe/layer.py
@@ -317,9 +317,25 @@ class MixtureOfExperts(keras.layers.Layer):
         for expert_id in range(self.num_experts):
             expert_output = self.experts[expert_id](inputs_flat, training=training)
 
-            # Weight and mask this expert's output
-            expert_weight = weights_flat[:, expert_id:expert_id + 1]
-            expert_mask = expert_assignment[:, expert_id:expert_id + 1]
+            # Weight and mask this expert's output.
+            #
+            # DECISION plan-2026-08-19T163559-499b6f0e/D-064
+            # Both factors are cast to the EXPERT OUTPUT's dtype. `ops.one_hot`
+            # returns float32 regardless of the active policy, so
+            # `expert_assignment` is float32 while the expert output under
+            # `mixed_float16` is float16, and the product raised
+            # `InvalidArgumentError: cannot compute Mul as input #1 was
+            # expected to be a half tensor but is a float tensor` on ANY
+            # mixed-precision forward through a hard-routed MoE -- MEASURED at
+            # HEAD via `Qwen3Next`, whose float32 control was green. Casting
+            # the gate rather than the expert output is deliberate: the routing
+            # weights are a convex combination in [0, 1], so half precision
+            # costs nothing there. See decisions.md D-064.
+            expert_weight = ops.cast(
+                weights_flat[:, expert_id:expert_id + 1], expert_output.dtype)
+            expert_mask = ops.cast(
+                expert_assignment[:, expert_id:expert_id + 1],
+                expert_output.dtype)
             weighted_output = expert_output * expert_weight * expert_mask
             expert_outputs.append(weighted_output)
